Add PYTHON_AFL_TSTL.
TSTL <https://github.com/agroce/tstl> is a domain-specific language and
set of tools to support automated generation of tests for software.

This change allows an environment variable (PYTHON_AFL_TSTL) to be set
causing the instrumentation to ignore code in the file "sut.py",
which contains a TSTL harness, not actual code-under-test. (Sometimes it
can be useful to include the harness code in the instrumentation, but
usually not.)